chore: clear the last two forge lint warnings in src/
`forge lint` reported 128 warnings; 126 are false positives for this codebase
(guarded reentrancy, precompile calls, zero-initialised loop counters, partial
tuple destructuring). These are the two that were not.

Auditing the existing suppressions by stripping all 15 `forge-lint: disable`
comments and diffing revealed only 9 warnings, all `unsafe-typecast`. Both
`block-timestamp` comments were dead: that lint flags comparisons, not reads,
so neither the read at `_startRecovery` nor the cached local in `syncBacking`
was ever a lint site. The live comparison that gates the write-off had no
suppression at all.

- Widen `Recovery.shortSince` to `uint256`. The `uint64` never packed, since
  the adjacent `parkedAtNonce` needs a full slot, so the narrowing bought no
  storage while forcing a cast and a mask on every read. Dropping it removes
  the `unsafe-typecast` site at its root rather than silencing it, and gas
  falls on every snapshotted path (-14 to -3700, no regressions).
- Suppress `block-timestamp` on the write-off comparison itself, with the
  reasoning: drift is seconds against a window of days, and an expired window
  only permits a write-off, which destroys holder value rather than extracting
  it. Delete the two comments that never applied, and the local that existed
  only to host one.
- Document that `consolidateRotatedStake` cannot return true under
  `RevertBelowFloor`, which is why its three callers discard the return.

`recovery()` is a public getter, so its ABI changes from (uint64,uint256) to
(uint256,uint256). Pre-launch, no deployment affected.

Four further dead suppressions remain in test/ (`AlphaVaultTestBase.sol:399,404`,
`mocks/MockStaking.sol:65`, `helpers/RegistryTestHelper.sol:33`) - forge-lint
does not apply `unsafe-typecast` to test files. Left in place: inert, accurate
about the cast, and live again if that policy changes.

Also ignore `.claude/`, which CLAUDE.md treats as never-commit but the repo did
not enforce.

+++ src/AlphaVaultLens.sol
@@ -74,7 +74,7 @@ contract AlphaVaultLens {
     /// @return deadline Recorded deadline, max uint256 for an undeclared shortfall, or zero if intact.
     /// @dev Expiry alone does not clear the loss.
     function writeOffDeadline(uint256 tokenId) external view returns (uint256 deadline) {
-        uint64 shortSince = _shortSince(tokenId);
+        uint256 shortSince = _shortSince(tokenId);
         if (shortSince != 0) return shortSince + vault.recoveryWindow();
         if (VaultReads.firstShortOf(_readBacking(tokenId).backing.short) != VaultReads.NO_SHORT_SLOT) {
             deadline = VaultReads.UNDECLARED_SHORTFALL;
@@ -171,7 +171,7 @@ contract AlphaVaultLens {
         return VaultMath.toNativeQuantum(VaultMath.backedEntitlement(entitlement, backing));
     }
 
-    function _shortSince(uint256 tokenId) private view returns (uint64 shortSince) { (shortSince,) = vault.recovery(tokenId); }
+    function _shortSince(uint256 tokenId) private view returns (uint256 shortSince) { (shortSince,) = vault.recovery(tokenId); }
 
     function _requireCurrentRegistration(uint256 tokenId) private view {
         if (VaultReads.isDissolved(tokenId)) revert SubnetDissolved();
